+++ src/__tests__/plopFilterBy.test.ts
@@ -1,5 +1,5 @@
 import test from 'ava'
-import { plopFilterBy } from '../plopFilterBy'
+import { plopFilterBy } from '../plop/plopFilterBy'
 
 test('validate filterBy collection', (t) => {
     const filterBy = plopFilterBy(`collection=My Collection`)
